+++ BookCorners/BookCorners/Views/Submit/SubmitLibraryView.swift
@@ -222,6 +222,8 @@ struct SubmitLibraryView: View {
                 ))
             }
 
+            OpenStreetMapContributionNotice()
+
             // MARK: - Submit
 
             Section {